pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ticket-godbattle"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "webtypy>=0.1.7",
    "pillow",
    "discord.py",
    "python-dotenv"
]

[tool.wrangler]
compatibility_date = "2026-02-22"
